backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
asyncpg==0.29.0
celery==5.3.6
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
openai==1.12.0
anthropic==0.18.1
pgvector==0.2.5
tenacity==8.2.3
email-validator==2.1.0.post1
